Simplification of single node benchmarks
Simplifies the framework for single node benchmarks. The overall
framework had gotten a little bit hacky and was difficult to maintain
and extend, so I decided to refactor. The focus is on easily building
and executing an OSB test against custom code. Details around what tests
to run and how to run them will be left in other repos.

Signed-off-by: John Mazanec <jmazane@amazon.com>